+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Add the `after_promotion` callback.
+- Removed the `before_fork` callback.
 - Fork workers and molds with a clean stack to allow more generations. (#30)
 
 # 0.1.2