+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- Remove default middlewares.
 - Refork indefinitely when `refork_after` is set, unless the last element is `false`.
 - Remove `mold_selector`. The promotion logic has been moved inside workers (#38).
 - Add the `after_promotion` callback.