+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 0.2.0
+
 - Remove default middlewares.
 - Refork indefinitely when `refork_after` is set, unless the last element is `false`.
 - Remove `mold_selector`. The promotion logic has been moved inside workers (#38).